+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 3 Dec 2025
+
+- Add server scheduler.
+
 # 2 Dec 2025
 
 - Add leaderboards.